Relax bound on DataBufRead for Cursor
This bound is overly restrictive - the point of Cursor<R> is to provide a Read implementation for any byte buffer.

Thus, it doesn't make sense to require that buffer itself also implements Read, as it prevents usage of Cursor with types like Vec<u8>, Rc<[u8]>, memmap2::Map and others.